fix: Resolve TypeScript compilation errors
Fixes type definition issues that were causing frontend test failures:

- Fix FloatingWindowData type import in state.ts
- Add ./scripts/app.js module declaration in app.d.ts
- Export FloatingWindowData from types.ts for re-export
- Fix ssh-dialog.ts imports (SSH functions from api/ssh.ts)
- Update dock.ts to import FloatingWindowData from state.ts

All TypeScript type errors resolved, 357 tests passing.
